Keep extract-candidates tests easier to extend
The extract-candidates suite repeated its extraction assertion helper inline, so this moves that shared expectation into a sibling helper module without changing Tailwind candidate coverage or edge-case assertions.

Constraint: Refactor must preserve extract-candidates test behavior and class extraction assertions exactly
Constraint: veryfront-code pre-commit verify:quick currently fails on pre-existing CLI boundary violations in cli/commands/extension/*.ts
Rejected: Collapse the suite into shared table-driven fixtures | would broaden the change and make the individual extraction cases harder to scan during the refactor wave
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Keep repeated extract-candidates test assertions in the helper module unless a future case needs bespoke inline inspection of the raw candidate list
Tested: deno test --no-check --allow-all tests/unit/html/styles-builder/extract-candidates.test.ts; deno fmt --check tests/unit/html/styles-builder/extract-candidates.test.ts tests/unit/html/styles-builder/extract-candidates.test-helpers.ts; deno lint tests/unit/html/styles-builder/extract-candidates.test.ts tests/unit/html/styles-builder/extract-candidates.test-helpers.ts; git diff --check
Not-tested: Full veryfront-code repo suite
